==> preprocess/1.create_title_dict.ipynb <==
import pandas as pd


df = pd.read_csv('../data/1.data_with_title.csv')

# df

translate_list = df[['title','short_tile']].to_dict(orient='index').values()

d = {}

for word in translate_list:
  d[word['title'].strip()] = word['short_tile']

d

import json

with open('../dictionary/title_dict.json','w',encoding='utf8')as file:
  json.dump(d,file,ensure_ascii=False)
