==> preprocess/4.classify_unit.ipynb <==
import json

with open('../dictionary/ingredient_dict_reqex.json','r') as file:
    ingredient_unit_dict = json.load(file)

import pandas as pd

df = pd.DataFrame(ingredient_unit_dict.items(), columns=['ingredient', 'unit'])


df

import sys
from pathlib import Path

# in jupyter (lab / notebook), based on notebook path
module_path = str(Path.cwd().parents[0])

if module_path not in sys.path:
    sys.path.append(module_path)

from dictionary.regex_util import tea_spoon, spoon, table_spoon, ladle , kilogram, gram, milligram, litre,glass, bottle, hundred_gram, hand, pick, lobe, rhizome, oz, cup, dipper, lid, min_unit, avg, buble,plate,millitre,drops

unit_dict = {
  'tea_spoon': tea_spoon,
  'spoon': spoon,
  'table_spoon': table_spoon,
  'ladle': ladle,
  'kilogram': kilogram,
  'gram': gram,
  'milligram': milligram,
  'litre': litre,
  'glass': glass,
  'bottle': bottle,
  'hundred_gram': hundred_gram,
  'hand': hand,
  'pick': pick,
  'lobe': lobe,
  'rhizome': rhizome,
  'oz': oz,
  'cup': cup,
  'dipper': dipper,
  'lid': lid,
  'min': min_unit,
  'avg': avg,
  'buble': buble,
  'plate': plate,
  'millitre': millitre,
  'drops': drops
}

import re

def is_any_match_regex(x,pattern,name):
  if(any([re.search(pattern, s) for s in x])):
    return name
  else: return ''
  

df['uniq_unit'] = df['unit']

def uniq_unit_by_regex(x: list[str],pattern,name):
  is_match = False
  new_x = []
  for s in x:
    if re.search(pattern,s.lower()):
      is_match = True
      continue
    else:
      new_x.append(s)
  if is_match:
    new_x.append(name)
  return new_x

for key,value in unit_dict.items():
  df[key] = df['unit'].apply(lambda x: is_any_match_regex(x,value,key))
  df['uniq_unit'] = df['uniq_unit'].apply(lambda x: uniq_unit_by_regex(x,value,key))

df

df[['uniq_unit']].sample(15)

df['not_amoung_unit'] = df['uniq_unit'].apply(lambda x : [s for s in x if s not in unit_dict])

df[['not_amoung_unit']]

export_df = df[['ingredient','not_amoung_unit']]

with open('../data/not_amoung_ingredient_unit.json', 'w', encoding='utf-8') as file:
    export_df.to_json(file,orient='records', force_ascii=False)

from dictionary.regex_util import tea_spoon,spoon,ladle,kilogram,plate,gram,milligram,millitre,litre,glass,bottle,hundred_gram,hand,pick,lid,lobe,rhizome,oz,cup,dipper,buble,min_unit,avg,drops

unit_dict = {
  tea_spoon: 'tea_spoon',
  spoon: 'spoon',
  table_spoon: 'table_spoon',
  ladle: 'ladle',
  kilogram: 'kilogram',
  plate: 'plate',
  gram: 'gram',
  milligram: 'milligram',
  millitre: 'millitre',
  litre: 'litre',
  glass: 'glass',
  bottle: 'bottle',
  hundred_gram: 'hundred_gram',
  hand: 'hand',
  pick: 'pick',
  lid: 'lid',
  lobe: 'lobe',
  rhizome: 'rhizome',
  oz: 'oz',
  cup: 'cup',
  dipper: 'dipper',
  buble: 'buble',
  drops: 'drops',
}

ramdom_unit = {
  min_unit: 'min_unit',
  avg: 'avg',
}

with open('../dictionary/unit_dict.json','r') as file:
    unit_dict_predict_amount = json.load(file)

unit_dict = { **unit_dict, **unit_dict_predict_amount, **ramdom_unit}

unit_dict

import json

with open('../dictionary/unit_full_dict.json','w',encoding='utf8')as file:
  json.dump(unit_dict,file,ensure_ascii=False)

import json

with open('../dictionary/unit_full_dict.json','w',encoding='utf8')as file:
  json.dump(unit_dict,file,ensure_ascii=False)
