==> preprocess/5.clean_ingredient_unit.ipynb <==
import pandas as pd

df = pd.read_json('../data/2.data_with_unit.json')

df

import sys
from pathlib import Path

# in jupyter (lab / notebook), based on notebook path
module_path = str(Path.cwd().parents[0])

if module_path not in sys.path:
    sys.path.append(module_path)

from dictionary.ingredient_dict import ingredient_dict

from dictionary.unit_dict import unit_dict, find_value_unit_dict

from pythainlp.util import normalize

import re
def map_ingredients_value(ingredients):
  fix_unit = False
  ingredients['ingredientName'] = normalize(ingredients['ingredientName'])
  for pattern in ingredient_dict:
    if re.search(pattern,ingredients['ingredientName']):
      ingredients['ingredientName'] = ingredient_dict[pattern]
      continue
  for pattern in unit_dict:
    if re.search(pattern,ingredients['unit']):
      ingredients['unit'] = unit_dict[pattern]
      fix_unit = True
      continue

  if not fix_unit:
    for pattern in find_value_unit_dict:
      if re.search(pattern,ingredients['unit']):
        if ingredients['amount'] != 0:
          print(ingredients)
        ingredients['unit'] = find_value_unit_dict[pattern] 
        if find_value_unit_dict[pattern] == 'min_unit':
          ingredients['amount'] = -2
        elif find_value_unit_dict[pattern] == 'avg':
          ingredients['amount'] = -1
  return ingredients

df['ingredients_generic'] = df['ingredients'].apply(lambda x: [map_ingredients_value(ingredient) for ingredient in x if ingredient['unit'] or ingredient['amount'] ])

df['ingredients'] = df.pop('ingredients_generic')
df

from pandas import json_normalize

df_normalized = json_normalize(df.sample()['ingredients'].tolist()[0])

df_normalized

with open('../data/final_ingredient_data.json', 'w', encoding='utf-8') as file:
    df.to_json(file,orient='records', force_ascii=False)

df_2 = pd.read_csv('../data/5.not_amoung_ingredient_unit_map_final.csv').reset_index()

df_2['ingredient'] = df_2['ingredient'].apply(lambda x: ingredient_dict[x])

df_2

ingre_name = df_2['ingredient'].tolist()

amount_unit_dict = {}

for x in df_2['ingredient']:
  amount_unit_dict[x] ={}

df_2 = df_2.drop(['ingredient','uniq_unit','Unnamed: 0','index'],axis=1)

df_2

for x in df_2:
  for y in range(len(df_2)):
    if df_2[x][y] != 0:
      try:
        amount_unit_dict[ingre_name[y]][x] = float(df_2[x][y])
      except:
        pass

amount_unit_dict

amount_unit_dict2 = {}
for k in amount_unit_dict:
  amount_unit_dict2[k] = {}
  for k2 in amount_unit_dict[k]:
    if amount_unit_dict[k][k2] != 0:
      amount_unit_dict2[k][k2] = amount_unit_dict[k][k2]

amount_unit_dict2

import json

with open('../dictionary/amount_ingredict.json','w',encoding='utf8')as file:
  json.dump(amount_unit_dict2,file,ensure_ascii=False)
