==> preprocess/8.reduce_ingredient.ipynb <==
import sys
from pathlib import Path

# in jupyter (lab / notebook), based on notebook path
module_path = str(Path.cwd().parents[0])

if module_path not in sys.path:
    sys.path.append(module_path)

from dictionary.ingredient_dict import ingredient_dict

ingredient_value = list(ingredient_dict.values())

import re

reduce_ingre = {}

def filter_By_word(d,word):
  return [x for x in  d if re.search(word,x)],[x for x in  d if not re.search(word,x)]

def reduce_ingredient_by_word(reduce_ingre,word,name,test=True):
  global ingredient_value

  a,b = filter_By_word(ingredient_value,word)
  print(a)
  if not test:
    reduce_ingre[name] = a
    ingredient_value = b

  return reduce_ingre

reduce_ingre = reduce_ingredient_by_word(reduce_ingre,'^(?!(น้ำจิ้ม|มัน)).*หมู|โบโลน่า|แฮม|เบคอน|เบค่อน|กระดูกอ่อน','หมู',test=False)

reduce_ingre = reduce_ingredient_by_word(reduce_ingre,'เครื่องใน','เครื่องใน',test=False)

reduce_ingre = reduce_ingredient_by_word(reduce_ingre,'^เนื้อ$','เนื้อ',test=False)

reduce_ingre = reduce_ingredient_by_word(reduce_ingre,'ไส้อั่ว|ไส้กรอก|กุนเชียง|แหนม|ฮอตดอก','เนื้อสัตว์แปรรูป',test=False)


reduce_ingre = reduce_ingredient_by_word(reduce_ingre,'หมึก','หมึก',test=False)

reduce_ingre = reduce_ingredient_by_word(reduce_ingre,'^(?!น้ำ).*ปลา','ปลา',test=False)

reduce_ingre = reduce_ingredient_by_word(reduce_ingre,'ไก่','ไก่',test=False)

reduce_ingre = reduce_ingredient_by_word(reduce_ingre,'กุ้ง','กุ้ง',test=False)

reduce_ingre = reduce_ingredient_by_word(reduce_ingre,'หอย','หอย')

reduce_ingre = reduce_ingredient_by_word(reduce_ingre,'ปู','ปู')

with open('../data/8.ingre.txt', 'w') as f:
  f.write('\n'.join(sorted(ingredient_value)))
  f.close()

with open('../data/8.ingre.txt', 'w') as f:
  f.write('\n'.join(sorted(ingredient_value)))
  f.close()

import json

with open('../dictionary/meat_ingredient.json','w',encoding='utf8')as file:
  json.dump(reduce_ingre,file,ensure_ascii=False)

reduce_ingredient_by_word('ซอส','ซอส',test=False)
