==> preprocess/9.merge_ingredient.ipynb <==
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)

import pandas as pd

df = pd.read_json('../data/final_ingredient_data.json')

import sys
from pathlib import Path

# in jupyter (lab / notebook), based on notebook path
module_path = str(Path.cwd().parents[0])

if module_path not in sys.path:
    sys.path.append(module_path)

import re
from dictionary.regex_util import number_with_special_character_regex
from dictionary.protein_carbo import protein_carbo_list
from dictionary.ingredient_merg import ingredient_merg
from dictionary.calories_list import calories_list
from dictionary.ingredient_dict import ingredient_dict
from dictionary.unit_dict import measure, amount_for_uncount_unit

def set_amount(ingredient):
  if ingredient['unit'] in measure:
    return ingredient['amount']*measure[ingredient['unit']] 
  elif ingredient['unit'] in amount_for_uncount_unit[ingredient['ingredientName']]:
    return ingredient['amount']*amount_for_uncount_unit[ingredient['ingredientName']][ingredient['unit']]
  if ingredient['amount'] > 0:
    return ingredient['amount']
  else: return 0

def merge_food_ingredient(ingredients):
  new_ingre = {}
  for ingre in ingredients:
    for m in ingredient_merg:
      if ingre['ingredientName'] in ingredient_merg[m]:
        if m not in new_ingre:
          new_ingre[m] = 0
        new_ingre[m]+= set_amount(ingre)
  return new_ingre

df['ingredients'][0]

df['mergeIngredients'] = df['ingredients'].apply(merge_food_ingredient)

def calulate_ingredient_cal(ingredients):
  cal = 0
  for ingre in ingredients:
    if ingre in calories_list:
      cal += calories_list[ingre]* ingredients[ingre]
  return cal

df['calories'] = df['mergeIngredients'].apply(calulate_ingredient_cal)

df[df['calories'] > 1500]

df.iloc[2719]['url']

df_yield = pd.read_json('../data/0.recipe_data_ugc.json')

def remove_leading_zeros(s: str):
    parts = s.split('.')
    if len(parts) == 2:
        left, right = parts
        if left == '0':
            left = ''
        return f"{left}.{right}"
    else:
        return s.lstrip('0') or '0'

def recur_change_to_number(number_list):
  if not number_list:
    return 0
  if '-' in number_list:
    dash_index = number_list.index('-')
    return (recur_change_to_number(number_list[:dash_index]) + recur_change_to_number(number_list[dash_index+1:]))/2
  number_list2 = []
  for num in number_list:
    try:
      number_list2.append(remove_leading_zeros(str(eval(num))))
    except:
      number_list2.append(num.strip())
  try:
    return eval(''.join(number_list2))
  except:
    # print(number_list2)


    if '+' == number_list2[-1]:
      recur_change_to_number(number_list2[:-1])
    else:
      # print('+'.join(number_list2))
      return eval('+'.join(number_list2))

def check_recur(num_list):
  a = recur_change_to_number(num_list)
  if a:
    return a
  return 1


df_yield['yield_num'] = df_yield['yield'].apply(lambda x: check_recur(re.findall(number_with_special_character_regex,x)))

df_yield['url'] = df_yield['url'].apply(lambda x: 'https://www.wongnai.com/'+ x)

df_yield = df_yield[['url','yield_num']]

merged_df = pd.merge(df, df_yield, on='url')

merged_df['calories2'] = merged_df['calories']/ merged_df['yield_num']

def merge_food_ingredient(ingredients):
  new_ingre = {}
  for ingre in ingredients:
    for m in ingredient_merg:
      if ingre['ingredientName'] in ingredient_merg[m]:
        if m not in new_ingre:
          new_ingre[m] = 0
        new_ingre[m]+= set_amount(ingre)
  return new_ingre

def merge_food_nutrients(ingredients):
  new_ingre = {}
  for ingre in ingredients:
    for m in protein_carbo_list:
      if ingre in protein_carbo_list[m]:
        if m not in new_ingre:
          new_ingre[m] = 0
        new_ingre[m]+= ingredients[ingre]
  return new_ingre

def cal_nutrients_percen(di):
  d = {**di}
  sumary = sum(d.values())
  if sumary:
    for x in d:
      d[x] = d[x]/sumary
  return d

merged_df['nutrients'] = merged_df['mergeIngredients'].apply(merge_food_nutrients)
merged_df['sum_nutrients'] = merged_df['nutrients'].apply(lambda x: sum(x.values()))
merged_df['nutrients_percen'] = merged_df['nutrients'].apply(cal_nutrients_percen)

merged_df.iloc[1]['ingredients']

def cal_spicy(x):
  if 'spicy_flavor' in x:
    return 3
  if 'mediumspicy_flavor' in x:
    return 2
  if 'little_spicy' in x:
    return 1
  return 0

merged_df['spicyLevel'] = merged_df['mergeIngredients'].apply(cal_spicy)

merged_df[merged_df['spicyLevel'] ==1].sample(3)

merged_df['calories'] = merged_df['calories2']
merged_df.drop('calories2',axis=1)

db_df = merged_df.copy(deep=False)

db_df = db_df[['url','calories','spicyLevel']]

with open('../data/addition_data.json','w',encoding='utf8')as file:
  db_df.to_json(file,orient='records', force_ascii=False)

merged_df[['title','tags','mergeIngredients','nutrients_percen','spicyLevel']]

model_df = merged_df[['title','tags','mergeIngredients','nutrients_percen','spicyLevel']]

from dictionary.tag_category import tag_category

tc = tag_category
del tc['ประเภทอาหาร']

all_tag = sum(list(tc.values()),[])

def is_external_id_in_list(external_ids, check_external_id):
  if(check_external_id in external_ids):
    return 1
  else:
    return 0

for external_id in all_tag:
  model_df[external_id] = model_df['tags'].apply(lambda tags : is_external_id_in_list([tag['externalId'] for tag in tags] ,external_id))

for ingre in ingredient_merg:
  model_df[ingre] = model_df['mergeIngredients'].apply(lambda mergeIngredients : is_external_id_in_list(list(mergeIngredients.keys()) ,ingre))

def is_in_dict(n,k):
  if k in n:
    return n[k]
  else :
    return 0

for protein_carbo in protein_carbo_list:
  model_df[protein_carbo] = model_df['nutrients_percen'].apply(lambda nutrients : is_in_dict(nutrients,protein_carbo))

model_df = model_df.drop(['tags','mergeIngredients','nutrients_percen'],axis=1)

model_df.to_csv('../data/data_for_model2.csv')
